+++ src/entities/User/types.ts
@@ -1,0 +1,8 @@
+import type { TMeta } from '../../shared/api';
+
+export interface IUser {
+	_id: TMeta['_id'];
+	firstName: string;
+	username: string;
+	avatar?: string;
+}
