+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './entities/Note';
+export * from './entities/User';
 
 export * from './shared/api';
